wth - (SPARCDashboard) Your Cost bad data
When a user changes the cost of a line item within admin edit, certain
entered strings are causing the page to error out and production bug
emails to be sent. Random strings are being filtered out ('Random
string' will return an error) but entering a partial decimal ('72.')
will cause it to error out.

I added a regular expression to check if the argument will be a valid
when running BigDecimal(argument), if it isn't, I convert it to an
integer first. [#145657373]